requirements.txt
streamlit
spacy==3.7.5
https://github.com/megagonlabs/ginza/releases/download/v5.2.0/ja_ginza-5.2.0-py3-none-any.whl
langchain
langchain-openai
langchain-anthropic
python-dotenv
fastapi
uvicorn
sqlalchemy